backend/market_algo_x/get_tickers.py
"""
FinancialModelingPrep API Stock Screener を使用して純粋な個別銘柄を取得

yfinanceの代わりにFMP Stock Screener APIを使用することで：
- ETF/投資信託を自動的に除外（isEtf=false, isFund=false）
- 複雑な文字列フィルタリングが不要
- 処理時間を数分〜十数分から数秒に短縮
- より正確な銘柄分類

環境変数 FMP_API_KEY が必要です。
"""

import os
import pandas as pd
from typing import List, Dict
import time
from curl_cffi.requests import Session
from dotenv import load_dotenv

# .envファイルから環境変数を読み込む
load_dotenv()


class FMPTickerFetcher:
    """FMP Stock Screener API を使用してティッカーを取得"""

    BASE_URL = "https://financialmodelingprep.com/api/v3/stock-screener"

    def __init__(self, api_key: str = None, rate_limit: int = None):
        """
        FMP Ticker Fetcherの初期化

        Args:
            api_key: FMP API Key（環境変数 FMP_API_KEY から自動取得可能）
            rate_limit: 1分あたりのAPIレート制限（環境変数 FMP_RATE_LIMIT から自動取得可能）
        """
        self.api_key = api_key or os.getenv('FMP_API_KEY')

        # We don't want to crash if env var is missing during initialization for other purposes,
        # but warn if it's missing.
        if not self.api_key:
            print("Warning: FMP_API_KEY is not set. Ticker fetching will fail.")

        # レート制限の設定（環境変数から取得、デフォルトは750 req/min - Premium Plan）
        self.rate_limit = rate_limit or int(os.getenv('FMP_RATE_LIMIT', '750'))
        self.session = Session(impersonate="chrome110")
        self.request_timestamps = []

    def _enforce_rate_limit(self):
        """設定されたAPIレート制限を適用"""
        current_time = time.time()
        # 60秒以上前のタイムスタンプを削除
        self.request_timestamps = [t for t in self.request_timestamps if current_time - t < 60]

        if len(self.request_timestamps) >= self.rate_limit:
            # 最も古いリクエストから60秒経過するまで待機
            sleep_time = 60 - (current_time - self.request_timestamps[0]) + 0.1
            # print(f"レート制限に達しました。{sleep_time:.1f}秒待機します...")
            time.sleep(sleep_time)
            # 待機後、再度古いタイムスタンプを削除
            current_time = time.time()
            self.request_timestamps = [t for t in self.request_timestamps if current_time - t < 60]

        self.request_timestamps.append(current_time)

    def _make_request(self, params: Dict) -> List[Dict]:
        """
        エラーハンドリングとレート制限を考慮したAPIリクエストを実行

        Args:
            params: クエリパラメータ

        Returns:
            List[Dict]: JSONレスポンス（辞書のリスト）
        """
        if not self.api_key:
             print("Error: FMP API Key is missing.")
             return []

        self._enforce_rate_limit()

        params['apikey'] = self.api_key

        try:
            response = self.session.get(self.BASE_URL, params=params)
            response.raise_for_status()
            data = response.json()

            if isinstance(data, list):
                return data
            elif isinstance(data, dict) and 'Error Message' in data:
                raise ValueError(f"API Error: {data['Error Message']}")
            else:
                # Some endpoints return dict, but stock-screener usually returns list.
                # If unexpected format, log it.
                if 'symbol' in data: # Single object?
                     return [data]
                print(f"Unexpected response format: {type(data)}")
                return []

        except Exception as e:
            print(f"API request failed: {e}")
            return []

    def get_stocks_by_exchange(self, exchange: str) -> List[Dict]:
        """
        指定された取引所から純粋な個別銘柄を取得

        Args:
            exchange: 取引所名 ('nasdaq', 'nyse', 'amex' など)

        Returns:
            銘柄情報のリスト
        """
        params = {
            'isEtf': 'false',              # ETF除外
            'isFund': 'false',             # 投資信託除外
            'isActivelyTrading': 'true',   # 取引停止中を除外
            'exchange': exchange.lower(),
            'limit': 10000                 # 最大取得数
        }

        # print(f"\n{exchange.upper()} の銘柄を取得中...")
        stocks = self._make_request(params)
        # print(f"  {len(stocks)} 件の銘柄を取得しました")

        return stocks

    def get_all_stocks(self, exchanges: List[str] = None) -> pd.DataFrame:
        """
        指定された取引所から全ての個別銘柄を取得

        Args:
            exchanges: 取引所のリスト（デフォルト: ['nasdaq', 'nyse', 'amex']）

        Returns:
            銘柄情報を含むDataFrame
        """
        if exchanges is None:
            exchanges = ['nasdaq', 'nyse', 'amex']

        all_stocks = []

        # Fallback if API Key is missing or invalid
        if not self.api_key or self.api_key == "your_fmp_api_key_here":
            print("FMP API Key is missing. Using S&P 500 fallback via Wikipedia...")
            return self._get_sp500_fallback()

        for exchange in exchanges:
            stocks = self.get_stocks_by_exchange(exchange)

            for stock in stocks:
                all_stocks.append({
                    'Ticker': stock.get('symbol'),
                    'Exchange': exchange.upper(),
                    'CompanyName': stock.get('companyName', ''),
                    'MarketCap': stock.get('marketCap', 0),
                    'Sector': stock.get('sector', ''),
                    'Industry': stock.get('industry', ''),
                    'Country': stock.get('country', '')
                })

        df = pd.DataFrame(all_stocks)

        if not df.empty:
            # 重複除外（同じティッカーが複数の取引所にリストされている場合）
            # 最初に見つかった取引所を優先
            df.drop_duplicates(subset=['Ticker'], keep='first', inplace=True)

        return df

    def _get_sp500_fallback(self) -> pd.DataFrame:
        """Fetch S&P 500 tickers from Wikipedia as a fallback"""
        try:
            import requests
            from bs4 import BeautifulSoup

            url = "https://en.wikipedia.org/wiki/List_of_S%26P_500_companies"
            response = requests.get(url)
            soup = BeautifulSoup(response.text, 'html.parser')
            table = soup.find('table', {'id': 'constituents'})

            tickers = []
            for row in table.find_all('tr')[1:]:
                cols = row.find_all('td')
                ticker = cols[0].text.strip().replace('.', '-')
                name = cols[1].text.strip()
                sector = cols[2].text.strip()
                sub_industry = cols[3].text.strip()

                tickers.append({
                    'Ticker': ticker,
                    'Exchange': 'NYSE', # Simplified assumption
                    'CompanyName': name,
                    'MarketCap': 0, # Not available
                    'Sector': sector,
                    'Industry': sub_industry,
                    'Country': 'USA'
                })

            # Limit to top 20 for faster testing if needed, or return all
            # Returning top 50 to ensure we get some hits
            return pd.DataFrame(tickers[:50])

        except Exception as e:
            print(f"Fallback fetch failed: {e}")
            # Absolute fallback
            return pd.DataFrame([
                {'Ticker': 'AAPL', 'Exchange': 'NASDAQ', 'Sector': 'Technology'},
                {'Ticker': 'MSFT', 'Exchange': 'NASDAQ', 'Sector': 'Technology'},
                {'Ticker': 'NVDA', 'Exchange': 'NASDAQ', 'Sector': 'Technology'},
                {'Ticker': 'AMZN', 'Exchange': 'NASDAQ', 'Sector': 'Consumer Cyclical'},
                {'Ticker': 'GOOGL', 'Exchange': 'NASDAQ', 'Sector': 'Communication Services'}
            ])
